backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
python-multipart==0.0.9
databricks-sql-connector==3.0.0
openai==1.58.1
